Add files via upload
Changed the sprite size from float to int

+++ render_sprite.c
@@ -6,7 +6,7 @@
 /*   By: aabounak <aabounak@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/20 05:24:27 by aabounak          #+#    #+#             */
-/*   Updated: 2020/12/05 18:12:42 by aabounak         ###   ########.fr       */
+/*   Updated: 2020/12/05 18:33:43 by aabounak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,20 +29,20 @@ void    render_sprite(int k, int x_fs, int y_fs)
 
 	init_spt_tex(k);
     i = 0;
-    while (i++ < (int)g_sprite[k]->size)
+    while (i++ < g_sprite[k]->size)
     {
         if (x_fs + i < 0 || x_fs + i > WIN_WIDTH)
             continue;
         if (g_rays[(int)(x_fs + i)].distance <= g_sprite[k]->distance)
             continue;
         j = 0;
-		while (j++ < (int)g_sprite[k]->size)
+		while (j++ < g_sprite[k]->size)
         {
             if (y_fs + j < 0 || y_fs + j > WIN_HEIGHT)
                 continue;
-            color = g_sprite[k]->data[(int)(g_sprite[k]->width * (j *
-					g_sprite[k]->height / (int)g_sprite[k]->size) +
-            		(i * g_sprite[k]->width / (int)g_sprite[k]->size))];
+            color = g_sprite[k]->data[g_sprite[k]->width * (j *
+					g_sprite[k]->height / g_sprite[k]->size) +
+            		(i * g_sprite[k]->width / g_sprite[k]->size)];
 			if (color)
 				img_update(x_fs + i, y_fs + j, color);
         }
@@ -71,7 +71,7 @@ void		ft_sprite_traits(int i)
 	printf("----------######----------\n");
 	printf("g_player.rotation_angle %f\n", g_player.rotation_angle);
 	printf("g_sprite[i]->angle %f\n", g_sprite[i]->angle);
-	printf("g_sprite[i]->size %f\n", g_sprite[i]->size);
+	printf("g_sprite[i]->size %d\n", g_sprite[i]->size);
 	printf("g_sprite[i]->distance %f\n", g_sprite[i]->distance);
 	printf("x_offset %f\n", x_offset);
 	printf("y_offset %f\n", y_offset);
